Changed Add Provider Function to Create New Directory

+++ Provider/Provider.py
@@ -1,5 +1,7 @@
 from Member import *
 import random
+import os
+from pathlib import Path
 
 from Service import *
 
@@ -25,19 +27,53 @@ class Provider:
         self.state= input("> ")
         print("Please enter the zipcode: ")
         self.zip = input("> ")
-        self.add_provider(self.provider_name,self.provider_id)
-        
+        self.add_provider()
+    
+    def print_exists(self):
+        print("Provider already exists.")  
     #test
-    def add_provider(self,name,id):
-        with open("Provider/ProviderList.json",mode="r") as file:
-            data = json.load(file)
-        new_provider = {
-            "ProviderName": name,
-            "ProviderId": id
-        }
-        data['providers'].append(new_provider)
-        with open("Provider/ProviderList.json",mode="w") as file:
-            json.dump(data,file,indent= 4)
+    def add_provider(self):
+        cwd = os.getcwd() #gets current working directory
+        parent_dir = "/ChocAn/Provider/" #sets relative path in variable
+        
+        #If provider already has a file, will print a statement that it exists
+        if os.path.exists(f"{cwd}/{parent_dir}/{self.provider_name}"): 
+            self.print_exists()
+        else:
+            #creates the new provider directory
+            directory = f"{self.provider_name}/" #new provider directory
+            p = Path(f"ChocAn/Provider/{directory}/") #sets path to the new provider's directory
+            p.mkdir() #makes the directory
+            #contents for file that is uploaded
+            provider = {
+                    "ProviderName": self.provider_name,
+                    "ProviderID": self.provider_id,
+                    "ProviderAddr": self.strAddr,
+                    "ProviderCity": self.city,
+                    "ProviderState": self.state,
+                    "ProviderZip": self.zip,
+                    "Services": [{
+                        
+                    }]
+            }
+            #Opens the path to the new folder and creates new json file for provider profile        
+            with open(f"{p}/{self.provider_name}.json",mode="w") as file:   #file 
+               json.dump(provider,file,indent= 4)
+
+            #Opens the path to the Provider folder for the
+            # list of providers in the json file
+            with open(f"{cwd}/{parent_dir}/ProviderList.json",mode="r") as file:
+                data = json.load(file)
+            #New provider data for the json file containing all providers in Provider folder
+            new_provider = {
+                "ProviderName": self.provider_name,
+                "ProviderId": self.provider_id
+            }
+            #Appends the new provider to the full provider list
+            data['providers'].append(new_provider)
+            with open(f"{cwd}/{parent_dir}/ProviderList.json",mode="w") as file:
+                json.dump(data,file,indent= 4)
+            
             
     #test
     def remove_provider(self):
